requirements.txt
cachetools==3.0.0
certifi==2018.11.29
chardet==3.0.4
defusedxml==0.5.0
Django==2.0.10
django-admin-sortable2==0.7
django-allauth==0.38.0
django-cleanup==3.0.1
django-jet==1.0.8
django-js-asset==1.1.0
django-mptt==0.9.1
django-smart-selects==1.5.4
django-solo==1.1.3
django-widget-tweaks==1.4.3
django-wysiwyg-redactor==0.5.1
google-api-python-client==1.7.7
google-auth==1.6.2
google-auth-httplib2==0.0.3
httplib2==0.12.0
idna==2.8
oauth2client==4.1.3
oauthlib==2.1.0
Pillow==5.4.1
pyasn1==0.4.5
pyasn1-modules==0.2.3
python-dateutil==2.7.5
python3-openid==3.1.0
pytz==2018.9
requests==2.21.0
requests-oauthlib==1.1.0
rsa==4.0
six==1.12.0
sorl-thumbnail==12.5.0
uritemplate==3.0.0
urllib3==1.24.1
